+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -100,10 +100,11 @@
     <Compile Include="EventWarlock\WarlockAction.cs" />
     <Compile Include="EventWarlock\WarlockActions\WarlockActionEcho.cs" />
     <Compile Include="EventWarlock\WarlockConditions\FirstEqualsSecond.cs" />
-    <Compile Include="MemoryDomains\MemoryDomainProxy.cs" />
-    <Compile Include="MemoryDomains\MemoryInterface.cs" />
-    <Compile Include="MemoryDomains\NullMemoryInterface.cs" />
-    <Compile Include="MemoryDomains\VirtualMemoryDomain.cs" />
+    <Compile Include="Memory\MemoryDomainProxy.cs" />
+    <Compile Include="Memory\MemoryDomains.cs" />
+    <Compile Include="Memory\MemoryInterface.cs" />
+    <Compile Include="Memory\NullMemoryInterface.cs" />
+    <Compile Include="Memory\VirtualMemoryDomain.cs" />
     <Compile Include="IRealTime.cs" />
     <Compile Include="ListFilters.cs" />
     <Compile Include="MemoryBanks.cs" />
