+++ Source/Libraries/CorruptCore/Memory/MemoryDomainProxy.cs
