+++ Source/Libraries/CorruptCore/Memory/MemoryInterface.cs
