+++ Source/Libraries/CorruptCore/Memory/NullMemoryInterface.cs
