+++ Source/Libraries/CorruptCore/Memory/VirtualMemoryDomain.cs
